% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_metre.R
\name{guess_metre}
\alias{guess_metre}
\title{Guess a poem's metre via Levenshtein distance from candidates}
\usage{
guess_metre(tokenized_poem)
}
\arguments{
\item{tokenized_poem}{}
}
\description{
Guess a poem's metre via Levenshtein distance from candidates
}
