% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_rhyme_type.R
\name{guess_rhyme_type}
\alias{guess_rhyme_type}
\title{Guess a poem's rhyme via Levenshtein distance from candidates}
\usage{
guess_rhyme_type(tokenized_poem)
}
\arguments{
\item{tokenized_poem}{}
}
\description{
Guess a poem's rhyme via Levenshtein distance from candidates
}
