% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhyme_sceme.R
\name{rhyme_scheme}
\alias{rhyme_scheme}
\title{Get a rhyme scheme for the poem.}
\usage{
rhyme_scheme(tokenized_poem)
}
\arguments{
\item{tokenized_poem}{}
}
\description{
For each line, lookahead to the future lines of the poem and see whether
last words rhyme.
}
