% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymes.R
\name{rhymes}
\alias{rhymes}
\title{For each word, get a list of various syllabic pronunications.}
\usage{
rhymes(word1, word2, level = 2)
}
\arguments{
\item{level}{}
}
\description{
Then check whether the last level number of syllables is pronounced
the same. If so, the words probably rhyme
}
