% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanscion.R
\name{scanscion}
\alias{scanscion}
\title{Get stress notation for every line in the poem.}
\usage{
scanscion(tokenized_poem)
}
\arguments{
\item{tokenized_poem}{}
}
\description{
Get stress notation for every line in the poem.
}
