% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanza_lengths.R
\name{stanza_lengths}
\alias{stanza_lengths}
\title{Returns a comma-delimited string of stanza lengths}
\usage{
stanza_lengths(tokenized_poem)
}
\arguments{
\item{tokenized_poem}{}
}
\description{
Returns a comma-delimited string of stanza lengths
}
